Wire up the pages to the select;
This patch adds the mail-in forms from the current `vote.gov` along with
some complex template logic to display the mail-in page. The data for
each registration form is handled within the Hugo data layer as well, so
adding new forms is relatively easy. Currently, this logic has a bias
for the English language and would need to be updated for multi-language
sites. Lastly, the JavaScript on the homepage template is currently a
placeholder in order to best demo the page layouts.