Create translated URLs from base content;
This patch adds the ability to generate new content and layout
directories related specifically to translating. These generated content
and layout directories are ignored by git and use only during the
development and build process. These files are necessary to create
translated URLs and should only be modified via gulp tasks. The current
`register/` directories in `content/` and `layouts/` are the base files
that are used to generate the Spanish files.