Update logic for language switcher;
This patch updates the logic used by the language switcher in order to
maintain state. This needs to be better documented as it's not easy to
understand at first glance. But, this works similarly to the way
language is being handled in the project and uses the
`data/translations/**/states.toml` configuration to determine which URL
to point to under which environment.